fit and predict function implemented
